move ServerSentEventManagerWithRetryPolicyTests test suite to new file
